Text index: embed min-max range into the index header
Currently, even if a token does not exists in the dictionary (assuming
bloom filter did not filter out), the dictionary block would be read
anyways to process the token.

To avoid it, the min-max range of the dictionary can be embedded into
text index header to filter out rare tokens. The storage overhead is
very minimal (two strings).